implement outlier detection LOF / DBSCAN on all 201808 detected peaks
